LayoutTests/fast/css-grid-layout/grid-item-spanning-resolution.html

<!DOCTYPE html>
<html>
<script>
if (window.testRunner)
    testRunner.overridePreference("WebKitCSSGridLayoutEnabled", 1);
</script>
<link href="resources/grid.css" rel="stylesheet">
<style>
.grid {
    -webkit-grid-definition-columns: 40% 60%;
    -webkit-grid-definition-rows: 30% 70%;
    width: 400px;
    height: 300px;
}

#bigGrid {
    -webkit-grid-definition-columns: 25% 25% 25% 25%;
    -webkit-grid-definition-rows: 25% 25% 25% 25%;
    height: 100px;
    width: 200px;
}

.negativeOverflowRowFirstColumn {
    -webkit-grid-row: 1 / -5;
    -webkit-grid-column: 1;
}

.overflowRowFirstColumn {
    -webkit-grid-row: 1 / 5;
    -webkit-grid-column: 1;
}

.firstRowNegativeOverflowColumn {
    -webkit-grid-row: 1;
    -webkit-grid-column: 1 / -5;
}

.firstRowOverflowColumn {
    -webkit-grid-row: 1;
    -webkit-grid-column: 1 / 5;
}

.secondRowSecondColumnNoSpan {
    -webkit-grid-column: 2 / 3;
    -webkit-grid-row: 2 / 3;
}

.thirdRowThirdColumnNoSpan {
    -webkit-grid-column: 3 / 4;
    -webkit-grid-row: 3 / 4;
}
</style>
<script src="../../resources/check-layout.js"></script>
<body onload="checkLayout('.grid');">

<p>Test that spannig rows / columns inside percentage sized grid areas get properly sized.</p>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea firstRowBothColumn" data-offset-x="0" data-offset-y="0" data-expected-width="400" data-expected-height="90"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea secondRowBothColumn" data-offset-x="0" data-offset-y="90" data-expected-width="400" data-expected-height="210"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea bothRowFirstColumn" data-offset-x="0" data-offset-y="0" data-expected-width="160" data-expected-height="300"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea bothRowSecondColumn" data-offset-x="160" data-offset-y="0" data-expected-width="240" data-expected-height="300"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea bothRowBothColumn" data-offset-x="0" data-offset-y="0" data-expected-width="400" data-expected-height="300"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea negativeOverflowRowFirstColumn" data-offset-x="0" data-offset-y="0" data-expected-width="160" data-expected-height="90"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea overflowRowFirstColumn" data-offset-x="0" data-offset-y="0" data-expected-width="160" data-expected-height="300"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea firstRowNegativeOverflowColumn" data-offset-x="0" data-offset-y="0" data-expected-width="160" data-expected-height="90"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea firstRowOverflowColumn" data-offset-x="0" data-offset-y="0" data-expected-width="400" data-expected-height="90"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea firstAutoRowSecondAutoColumn" data-offset-x="160" data-offset-y="0" data-expected-width="240" data-expected-height="90"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" data-expected-width="400" data-expected-height="300">
    <div class="sizedToGridArea autoSecondRowAutoFirstColumn" data-offset-x="0" data-offset-y="0" data-expected-width="160" data-expected-height="90"></div>
</div>
</div>

<div style="position: relative">
<div class="grid" id="bigGrid" data-expected-width="200" data-expected-height="100">
    <div class="sizedToGridArea secondRowSecondColumnNoSpan" data-offset-x="50" data-offset-y="25" data-expected-width="50" data-expected-height="25"></div>
    <div class="sizedToGridArea thirdRowThirdColumnNoSpan" data-offset-x="100" data-offset-y="50" data-expected-width="50" data-expected-height="25"></div>
</div>
</body>
</html>
